bump version to 1.19.16, change id fields from Long to Integer in Organization, Application, and ApplicationDao classes, and update related methods in ApplicationDaoIface and IotDatabaseDao

+++ src/main/java/com/signomix/common/Organization.java
@@ -1,13 +1,13 @@
 package com.signomix.common;
 
 public class Organization {
-    public Long id;
+    public Integer id;
     public String code;
     public String name;
     public String description;
     public Integer numberOfTenants;
 
-    public Organization(Long id, String code, String name, String description, Integer numberOfTenants) {
+    public Organization(Integer id, String code, String name, String description, Integer numberOfTenants) {
         this.id = id;
         this.name = name;
         this.code = code;
